service/requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; sys_platform == "win32" and python_version >= "3.10" and python_version < "4.0" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
elastic-transport==8.4.0 ; python_version >= "3.10" and python_version < "4"
elasticsearch==8.5.0 ; python_version >= "3.10" and python_version < "4"
fastapi==0.87.0 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.1 ; python_version >= "3.10" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy-utils==0.38.3 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.44 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy[asyncio]==1.4.44 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.10" and python_version < "4"
uvicorn[standard]==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.18.1 ; python_version >= "3.10" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.10" and python_version < "4.0"
